Pedantically restructure certain areas
- maps
- metadata
- predictions
