+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pytide"
-version = "0.3.1"
+version = "0.3.2"
 description = "Pytide: A tool for emailing high- and low-tide events."
 readme = "README.md"
 authors = [
