+++ src/pytide/metadata/repository.py
@@ -1,34 +1,9 @@
 import sqlite3
 
-import requests
-
 from pytide.database.cache import get_connection
-from pytide.metadata.models import FetchNoaaMetadataResponse, GetCachedMetadataResponse, SaveMetadataRequest
+from pytide.metadata.models import GetCachedMetadataResponse, SaveMetadataRequest
 
 CACHE_EXPIRATION = '-7 days'
-
-
-def fetch_noaa_metadata() -> list[FetchNoaaMetadataResponse]:
-    api_url = 'https://api.tidesandcurrents.noaa.gov/mdapi/prod/webapi/stations.json?type=tidepredictions'
-
-    try:
-        response = requests.get(api_url, timeout=10)
-        response.raise_for_status()
-
-        stations = response.json()['stations']
-
-        return [
-            FetchNoaaMetadataResponse(
-                station['id'],
-                station['name'],
-                round(station['lat'], 6),
-                round(station['lng'], 6),
-            )
-            for station in stations
-        ]
-
-    except requests.RequestException as error:
-        raise SystemExit(f'Unable to retrieve station metadata -> {error}') from error
 
 
 def cache_is_fresh() -> bool:
